fix: high quality curve setting is not preserved properly
This setting is set to _off_ on some imported animation files;
in this case, tangent data is not used and the curve is interpolated linearly.

Previously, we would sometimes set this to on as part of various animation
manipulations we perform in MA processing; this broke some avatars, e.g. "Fluria".