fix: Animator can cache incorrect humanoid bone mappings (#431)
When multiple bones with the same name exist in the initial hierarchy, the root Animator
can cache (prior to MA execution) bones that will later be deleted or renamed due to MA
action. While this isn't persisted to the avatar asset bundle, it can result in validation failures
blocking upload. To resolve this, invalidate this cache late in MA execution.