fix: avoid name collisions in setup outfit
The human avatar mapping system seems to use bone _names_ rather than full
_paths_ to identify bones. When the avatar armature and an outfit armature
are both present under the avatar root, this can result in misidentification
of outfit bones as avatar bones on the avatar animator. This in turn results
in issues with Bone Proxy's editor-side tracking logic.

This change adjusts setup outfit to ensure that there is always a prefix
and/or suffix set, renaming bones if necessary.

Note that this does not fully use outfit human avatar data to map bones yet;
this is mostly intended as a patch to resolve the issues that have been
reported recently, particularly around the stricter validations in SDK
3.3.0.