fix: NaNimation initial states are not applied when animations are inactive
This changes the NaNimation system to work using blendshapes (with
infinite deltaPositions) instead of scaled bones. While this has
slightly higher overhead in some cases, the bone technique cannot be
used in the initial (pre-animator) state of the avatar, while the
blendshape technique work in this case.

Additionally, the blendshape technique does not increase the bone count
of the avatar, thus allowing for more masks on Quest.

Closes: #1657