fix: NaNimation initial states are not applied when animations are inactive
This fixes #1657 in a way which does not trigger VRChat crashes in
certain worlds. Unlike #1660, which used infinite blendshape
displacement to erase polygons (which has side effects such as crashing,
and causing issues with orthographic cameras), we instead revert to
using NaNimation, and use VRC Scale Constraints to NaNimate the initial
state of our NaNimation bones.