fix: culling issues with rootbone=head
When we perform NaNimations, we disable the "update when offscreen" option on the affected renderer.
This has a side effect: If the root bone is a bone underneath the head, then when the head is scaled down to
a near-zero size, the bounding box of the renderer is also scaled down to a near-zero size, and is then culled
in first-person view.

To avoid this, we now detect when the root bone is the head (or a child of the head) and, in this case, create a
substitute root bone with VRCHeadChop to disable this behavior.