fix: default value could be set to zero for menu item params
When a menu item parameter is set to automatic value with isDefault ON, but the parameter is controlled by a parent MA Parameters with a default of zero, we would previously set the menu item value to zero. This breaks the menu item as it's not possible to turn this menu item "off", as it just sets the parameter back to zero.

This change adjusts this logic to ignore the MA Parameters default value if it is zero; this then means we'll assign values starting at 1.